Move ListQueueValueNotifier out of shared + add TODO for tests
