+++ packages/devtools_app/lib/src/utils/list_queue_value_notifier.dart
