+++ packages/devtools_app/test/shared/utils/list_queue_value_notifier_test.dart
@@ -1,0 +1,7 @@
+// Copyright 2024 The Chromium Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+void main() {
+  // TODO(dantup): Add tests for this + consider moving to devtools_app_shared.
+}
